+++ src/create/prefab_creator.py
@@ -19,16 +19,6 @@ from src.ecs.components.c_transform import CTransform
 from src.ecs.components.c_velocity import CVelocity
 from src.ecs.components.tags.c_tag_player import CTagPlayer
 from src.engine.service_locator import ServiceLocator
-    
-def create_square(ecs_world:esper.World, size:pygame.Vector2,
-                   pos:pygame.Vector2, vel:pygame.Vector2, col:pygame.Color) -> int:
-    cuad_entity = ecs_world.create_entity()
-    ecs_world.add_component(cuad_entity,
-                                    CSurface(size, col))
-    ecs_world.add_component(cuad_entity,
-                CTransform(pos))
-    ecs_world.add_component(cuad_entity, 
-                CVelocity(vel))
 from src.ecs.components.c_input_command import CInputCommand
 from src.ecs.components.c_star_field import CStarField
 from src.ecs.components.c_surface import CSurface
@@ -37,6 +27,17 @@ from src.ecs.components.c_velocity import CVelocity
 from src.ecs.components.tags.c_tag_bullet import CTagBullet
 from src.ecs.components.tags.c_tag_player import CTagPlayer
 from src.engine.service_locator import ServiceLocator
+
+def create_square(ecs_world:esper.World, size:pygame.Vector2,
+                   pos:pygame.Vector2, vel:pygame.Vector2, col:pygame.Color) -> int:
+    cuad_entity = ecs_world.create_entity()
+    ecs_world.add_component(cuad_entity,
+                                    CSurface(size, col))
+    ecs_world.add_component(cuad_entity,
+                                    CTransform(pos))
+    ecs_world.add_component(cuad_entity,
+                                    CVelocity(vel))
+    return cuad_entity
     
 def create_sprite(ecs_world:esper.World, pos:pygame.Vector2, vel:pygame.Vector2, 
                   surface:pygame.Vector2) -> int:
@@ -82,6 +83,20 @@ def create_input_player(ecs_world:esper.World):
     ecs_world.add_component(input_right, CInputCommand("PLAYER_RIGHT", [pygame.K_RIGHT, pygame.K_d]))
     ecs_world.add_component(input_space, CInputCommand("PLAYER_FIRE", [pygame.K_SPACE, pygame.K_z]))
     #ecs_world.add_component(input_p, CInputCommand("PLAYER_PAUSE", [pygame.K_p]))
+
+def create_star(ecs_world:esper.World, window_cfg, starfield_cfg):
+    window_width = window_cfg["size"]["w"]
+    window_height = window_cfg["size"]["h"]
+    num_stars = starfield_cfg["number_of_stars"]
+    for _ in range(num_stars):
+        pos = pygame.Vector2(random.randint(0, window_width),
+                             random.randint(0, window_height))
+        vel = pygame.Vector2(random.uniform(0, 0), random.uniform(starfield_cfg["vertical_speed"]["min"], starfield_cfg["vertical_speed"]["max"]))
+        star_color = random.choice(starfield_cfg["star_colors"])
+        star_surface = pygame.Surface((starfield_cfg["size"]["h"], starfield_cfg["size"]["w"]))
+        star_surface.fill((star_color["r"], star_color["g"], star_color["b"]))
+        star_entity = create_sprite(ecs_world, pos, vel, star_surface)
+        ecs_world.add_component(star_entity, CStarField(window_width, window_height, star_surface, starfield_cfg["blink_rate"]["min"], starfield_cfg["blink_rate"]["max"], star_color))
 
 def create_enemy(ecs_world:esper.World, position:pygame.Vector2, velocity:int,
                   enemy_info:dict, type:str):
