+++ src/ecs/components/c_star_field.py
@@ -1,4 +1,3 @@
-
 import random
 
 import pygame
